IGNITE-28970 Remove stale methods from OptimizedMarshaller
