+++ modules/binary/impl/src/main/java/org/apache/ignite/internal/marshaller/optimized/OptimizedMarshallerImpl.java
@@ -86,19 +86,16 @@ public class OptimizedMarshallerImpl extends AbstractNodeNameAwareMarshaller imp
     private final ClassLoader dfltClsLdr = getClass().getClassLoader();
 
     /** Whether or not to require an object to be serializable in order to be marshalled. */
-    private boolean requireSer = true;
-
-    /** ID mapper. */
-    private OptimizedMarshallerIdMapper mapper;
+    private final boolean requireSer;
 
     /** Class descriptors by class. */
     private final ConcurrentMap<Class, OptimizedClassDescriptor> clsMap = new ConcurrentHashMap<>();
 
     /** */
-    private OptimizedObjectStreamRegistry registry = new OptimizedObjectSharedStreamRegistry();
+    private final OptimizedObjectSharedStreamRegistry registry = new OptimizedObjectSharedStreamRegistry();
 
     /** Non cached registry. */
-    private OptimizedObjectSharedStreamRegistry nonCachedRegistry = new OptimizedObjectSharedStreamRegistry();
+    private final OptimizedObjectSharedStreamRegistry nonCachedRegistry = new OptimizedObjectSharedStreamRegistry();
 
     /**
      * Creates new marshaller will all defaults.
@@ -106,6 +103,8 @@ public class OptimizedMarshallerImpl extends AbstractNodeNameAwareMarshaller imp
      * @throws IgniteException If this marshaller is not supported on the current JVM.
      */
     public OptimizedMarshallerImpl() {
+        this.requireSer = true;
+
         if (!available())
             throw new IgniteException("Using OptimizedMarshaller on unsupported JVM version (some of " +
                 "JVM-private APIs required for the marshaller to work are missing).");
@@ -122,29 +121,6 @@ public class OptimizedMarshallerImpl extends AbstractNodeNameAwareMarshaller imp
     }
 
     /** {@inheritDoc} */
-    @Override public OptimizedMarshaller setRequireSerializable(boolean requireSer) {
-        this.requireSer = requireSer;
-
-        return this;
-    }
-
-    /** {@inheritDoc} */
-    @Override public OptimizedMarshaller setIdMapper(OptimizedMarshallerIdMapper mapper) {
-        this.mapper = mapper;
-
-        return this;
-    }
-
-    /** {@inheritDoc} */
-    @Override public OptimizedMarshaller setPoolSize(int poolSize) {
-        registry = poolSize > 0 ?
-            new OptimizedObjectPooledStreamRegistry(poolSize) :
-            new OptimizedObjectSharedStreamRegistry();
-
-        return this;
-    }
-
-    /** {@inheritDoc} */
     @Override protected void marshal0(@Nullable Object obj, OutputStream out) throws IgniteCheckedException {
         assert out != null;
 
@@ -153,7 +129,7 @@ public class OptimizedMarshallerImpl extends AbstractNodeNameAwareMarshaller imp
         try {
             objOut = registry.out();
 
-            objOut.context(clsMap, ctx, mapper, requireSer);
+            objOut.context(clsMap, ctx, null, requireSer);
 
             objOut.out().outputStream(out);
 
@@ -174,7 +150,7 @@ public class OptimizedMarshallerImpl extends AbstractNodeNameAwareMarshaller imp
         try {
             objOut = registry.out();
 
-            objOut.context(clsMap, ctx, mapper, requireSer);
+            objOut.context(clsMap, ctx, null, requireSer);
 
             objOut.writeObject(obj);
 
@@ -199,7 +175,7 @@ public class OptimizedMarshallerImpl extends AbstractNodeNameAwareMarshaller imp
         try {
             objIn = !useCache ? nonCachedRegistry.in() : registry.in();
 
-            objIn.context(clsMap, ctx, mapper, clsLdr != null ? clsLdr : dfltClsLdr);
+            objIn.context(clsMap, ctx, null, clsLdr != null ? clsLdr : dfltClsLdr);
 
             objIn.in().inputStream(in);
 
@@ -232,7 +208,7 @@ public class OptimizedMarshallerImpl extends AbstractNodeNameAwareMarshaller imp
         try {
             objIn = registry.in();
 
-            objIn.context(clsMap, ctx, mapper, clsLdr != null ? clsLdr : dfltClsLdr);
+            objIn.context(clsMap, ctx, null, clsLdr != null ? clsLdr : dfltClsLdr);
 
             objIn.in().bytes(arr, arr.length);
