+++ modules/core/src/test/java/org/apache/ignite/internal/marshaller/optimized/OptimizedMarshallerTest.java
@@ -53,9 +53,17 @@ public class OptimizedMarshallerTest extends GridCommonAbstractTest {
      * @return Marshaller.
      */
     private OptimizedMarshaller marshaller() {
+        return marshaller(true);
+    }
+
+
+    /**
+     * @return Marshaller.
+     */
+    private OptimizedMarshaller marshaller(boolean requireSer) {
         ClassLoaderUtils.clearClassCache();
 
-        OptimizedMarshaller marsh = Marshallers.optimized();
+        OptimizedMarshaller marsh = Marshallers.optimized(requireSer);
 
         marsh.setContext(new MarshallerContextTestImpl());
 
@@ -69,9 +77,7 @@ public class OptimizedMarshallerTest extends GridCommonAbstractTest {
      */
     @Test
     public void testNonSerializable() throws IgniteCheckedException {
-        OptimizedMarshaller marsh = marshaller();
-
-        marsh.setRequireSerializable(false);
+        OptimizedMarshaller marsh = marshaller(false);
 
         NonSerializable outObj = marsh.unmarshal(marsh.marshal(new NonSerializable(null)), null);
 
@@ -85,9 +91,7 @@ public class OptimizedMarshallerTest extends GridCommonAbstractTest {
      */
     @Test
     public void testNonSerializable1() throws IgniteCheckedException {
-        OptimizedMarshaller marsh = marshaller();
-
-        marsh.setRequireSerializable(false);
+        OptimizedMarshaller marsh = marshaller(false);
 
         byte[] bytes = marsh.marshal(new TcpDiscoveryVmIpFinder());
 
@@ -107,9 +111,7 @@ public class OptimizedMarshallerTest extends GridCommonAbstractTest {
      */
     @Test
     public void testNonSerializable2() throws IgniteCheckedException {
-        OptimizedMarshaller marsh = marshaller();
-
-        marsh.setRequireSerializable(false);
+        OptimizedMarshaller marsh = marshaller(false);
 
         TcpDiscoveryIpFinderAdapter ipFinder = new TcpDiscoveryIpFinderAdapter() {
             @Override public Collection<InetSocketAddress> getRegisteredAddresses() {
@@ -141,9 +143,7 @@ public class OptimizedMarshallerTest extends GridCommonAbstractTest {
      */
     @Test
     public void testNonSerializable3() throws IgniteCheckedException {
-        OptimizedMarshaller marsh = marshaller();
-
-        marsh.setRequireSerializable(false);
+        OptimizedMarshaller marsh = marshaller(false);
 
         byte[] bytes = marsh.marshal(new TestTcpDiscoveryIpFinderAdapter());
 
@@ -159,9 +159,7 @@ public class OptimizedMarshallerTest extends GridCommonAbstractTest {
      */
     @Test
     public void testNonSerializable4() throws IgniteCheckedException {
-        OptimizedMarshaller marsh = marshaller();
-
-        marsh.setRequireSerializable(false);
+        OptimizedMarshaller marsh = marshaller(false);
 
         byte[] bytes = marsh.marshal(new GridMarshallerTestInheritedBean());
 
@@ -238,14 +236,10 @@ public class OptimizedMarshallerTest extends GridCommonAbstractTest {
         assertNotNull(outObj1);
     }
 
-    /**
-     * Tests {@link OptimizedMarshaller#setRequireSerializable(boolean)}.
-     */
+    /** */
     @Test
     public void testRequireSerializable() {
         OptimizedMarshaller marsh = marshaller();
-
-        marsh.setRequireSerializable(true);
 
         try {
             marsh.marshal(new NonSerializable(null));
@@ -264,9 +258,7 @@ public class OptimizedMarshallerTest extends GridCommonAbstractTest {
      */
     @Test
     public void testProxy() throws IgniteCheckedException {
-        OptimizedMarshaller marsh = marshaller();
-
-        marsh.setRequireSerializable(false);
+        OptimizedMarshaller marsh = marshaller(false);
 
         SomeItf inItf = (SomeItf)Proxy.newProxyInstance(
             OptimizedMarshallerTest.class.getClassLoader(), new Class[] {SomeItf.class},
